multi-pack-index: use PARSE_OPT_ERROR_AT_NON_OPTION
Make use of the new PARSE_OPT_ERROR_AT_NON_OPTION flag in the
multi-pack-index code. This code added in
60ca94769ce (builtin/multi-pack-index.c: split sub-commands,
2021-03-30) had the unusual pattern of passing PARSE_OPT_KEEP_UNKNOWN,
and then proceeding to die if there was anything left over in argc.

It's unusual because by doing that we don't emit errors about unknown
options to these sub-commands, now we will:

    $ git multi-pack-index verify --unknown
    error: unknown option `unknown'
    usage: git multi-pack-index [<options>] verify
    [...]

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>