http: drop support for curl < 7.18.0
In a preceding commit we dropped support for curl < 7.17.0, let's
continue that and drop support for versions older than 7.18.0. This
allows us to get rid of some now-obsolete #ifdefs around
"CURLPROXY_SOCKS".

This change looks incorrect at first sight, because in curl's hex
version notation 0x071800 is version 7.24.0, *not* 7.18.0.

But it's correct, because the existing version check being removed
here is wrong. The check guards use of the following curl options,
introduced at these respective versions[1]:

    CURLPROXY_SOCKS4                7.10
    CURLPROXY_SOCKS4A               7.18.0
    CURLPROXY_SOCKS5                7.10
    CURLPROXY_SOCKS5_HOSTNAME       7.18.0

I.e. the oldest version that has these is in fact 7.18.0, not
7.24.0. That we were checking 7.24.0 is just an mistake in
6d7afe07f29 (remote-http(s): support SOCKS proxies, 2015-10-26) that
slipped past review. I.e. the version check confusing base 10 and with
base 16.

1. https://github.com/curl/curl/blob/master/docs/libcurl/symbols-in-versions

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>