protocol v2 tests: split up packfile-uri tests
The t5702-protocol-v2.sh file is starting to take quite a long time to
run, let's split up the packfile-uri tests into their own test
file. After this change the two run in around 5 and 2 seconds,
respectively.

I have more general patches to split up tests in this area even
further, in particular to address the issue of e.g. HTTPD tests not
being run on setups that want to skip "git-daemon" tests.

This change is just a small part of that, but sufficient for the
purposes of subsequent patches, which will extend the packfile-uri /
.gitmodules tests here.

See dd4b732df73 (upload-pack: send part of packfile response as uri,
2020-06-10) for the initial addition of these tests.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>