protocol v2 tests: use test_when_finished in packfile-uri tests
Change the packfile-uri tests to remove the things they themselves
generated, instead of relying on later tests to do so. This makes
adding more tests easier, as we'll be able to use test_when_finished
in those, without worrying about leftover files from other tests.

I'm omitting the removal of "err" here, since it's typically used in a
way that's always clobbered.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>