advice.[ch]: remove advice.pushNonFastForward alias
The advice.pushNonFastForward variable has been primarily known as
advice.pushUpdateRejected since 1184564eac8 (push: rename config
variable for more general use, 2012-12-02). It's been long enough by
far that we can declare this deprecation period over and remove it.

The only downside for users is going to be that they'll potentially
see a more verbose warning that they'll have to re-squash by changing
their config, but that's hardly an undue burden.

Doing this allows us to simplify the advice_enabled() function down to
something truly trivial.

Change the only in-tree user of this, it was added in
fd47d7b94db (git-remote-mediawiki: obey advice.pushNonFastForward,
2011-09-27), and not renamed in 1184564eac8 along with the rest.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>