advice API: invert always-true "enabled" field
Rename the "enabled" field which is always true by default (but can be
disabled by config) to "disabled". This allows us to omit the
boilerplate of initializing that field to "1" for the large static
"advice_setting" assignment.

Let's also write those variables with the "advice." prefix instead of
adding/skipping it dynamically. This makes it easier to grep for this
configuration, and as the diff shows saves us (a trivial amount of)
work at runtime.

Further, let's convert this to use designated initializers. Doing so
makes it immediately apparent that we're assigning to the "key"
member.

See b3b18d16213 (advice: revamp advise API, 2020-03-02) for the last
revamping of this API.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>