advice API: invert always-true "enabled" field
Rename the "enabled" field which is always true by default (but can be
disabled by config) to "disabled", and make it a bitfield.

This allows us to omit the boilerplate of initializing that field to
"1" for the large static "advice_setting" assignment.

This change was initially made before generate-advice.sh was created,
so the reduction in boilerplate are rather trivial at this point, but
this still simplifies and clarifies things a bit.

Let's also write those variables with the "advice." prefix instead of
adding/skipping it dynamically. This makes it easier to grep for this
configuration, and as the diff shows saves us (a trivial amount of)
work at runtime.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>