advice output: consistently emit "\n\n" before config instructions
The instructions we'd emit for showing how to disable advice messaging
was inconsistent, in some cases we'd emit it, in others not. Let's
always emit that empty line before the config instructions.

As we'll see in a subsequent commit there are some users of the
advise() API that didn't use this "display_instructions" codepath, but
rather emitter their own instructions, in the end these will all be
consistent.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>