advice API: always show config suggestion with advise()
Change the advise() output to emit instructions at the end about how
to disable the message we just emitted, using the new "type" parameter
to advise() added in the preceding commit.

This brings it in line with how advice_if_enabled() has worked since,
b3b18d16213 (advice: revamp advise API, 2020-03-02). That commit
suggested that it was a thin wrapper for advise(), but their output
was different, now it's the same.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>