strvec API: expose the strvec_push_nodup() function
Ever since this API was exposed for public use in
c1189caeaf7 (refactor argv_array into generic code, 2011-09-13) as
argv-array.h it's been documented that it'll always duplicate the
pushed arguments.

For almost all its users that's what we want, but for a few of them
it's needlessly inconvenient that you can't e.g. strbuf_detach() an
already allocated string. Let's support that use-case by exposing the
strvec_push_nodup() function.

I'm changing it from returning "void" to returning the same "const
char *" as strvec_push() does for consistency. That makes it trivial
to implement the two in terms of one another.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>