reftable: fixup for broken __FUNCTION__ use
The use of the __FUNCTION__ macro is non-standard, in this case all
we're doing with it is generating a prettier unique filename based on
the name of the function that makes it, presumably to make ad-hoc
debugging easier.

For mkdtemp() and mkstemp() we don't need to pass anything like this
in, since the "XXXXXX" part of the template will ensure that we get a
unique filename, but to make finding what function created what
tempfile easy let's just use __LINE__ here, it's not *as easy*, but at
least this one uses standard behavior.

This can be tested under DEVOPTS=pedantic, i.e. before this change
we'd emit errors like:

    reftable/stack_test.c: In function ‘test_read_file’:
    reftable/stack_test.c:67:30: error: ISO C does not support ‘__FUNCTION__’ predefined identifier [-Werror=pedantic]
      char *fn = get_tmp_template(__FUNCTION__);

The current tip of "seen" is broken as a result, see
https://github.com/git/git/runs/3439941236

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>