commit-graph: fix memory leak in misused string_list API
When this code was migrated to the string_list API in
d88b14b3fd6 (commit-graph: use string-list API for input, 2018-06-27)
it used both STRING_LIST_INIT_NODUP and a strbuf_detach() pattern,
we'd thus never free() the allocated strings. Let's do that, and clean
up the control flow and variable declarations related to that.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>