repack/show-index/write-tree: error on excess arguments
Use the new PARSE_OPT_ERROR_AT_NON_OPTION to parse_options() to have
"repack", "show-index" and "write-tree" error out whe given arguments
they're not handling, before they'd silently ignore them.

Let's also test for this, in the case of "show-index" and "write-tree"
there was no dedicated test file for them, but their functionality is
tested in various other tests. Let's add a start at a test file for
both, right now we'll just check this part of the expected usage, but
now we'll have a logical place to add more tests in the future.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>