pack-write: split up finish_tmp_packfile() function
Split up the finish_tmp_packfile() function and use the split-up
version in pack-objects.c in preparation for moving the step of
renaming the *.idx file later as part of a function change.

Since the only other caller of finish_tmp_packfile() was in
bulk-checkin.c, and it won't be needing a change to its *.idx
renaming, provide a thin wrapper for the old function as a static
function in that file. If other callers end up needing the simpler
version it could be moved back to "pack-write.c" and "pack.h".

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>